refactor: move algorithm internals to knn_algorithm.py; remove knncpd.cpd.py
