+++ CPDShell/Core/algorithms/knn_algorithm.py
@@ -9,10 +9,11 @@ __license__ = "SPDX-License-Identifier: MIT"
 import numpy as np
 import typing as tp
 
+from math import sqrt
 from collections import deque
 from collections.abc import Iterable
 
-from CPDShell.Core.algorithms.KNNCPD.cpd import KNNCPD
+import KNNCPD.knn_graph as knngraph
 from CPDShell.Core.algorithms.abstract_algorithm import Algorithm
 
 
@@ -24,10 +25,14 @@ class KNNAlgorithm(Algorithm):
     def __init__(
         self, metric: tp.Callable[[float, float], float] | tp.Callable[[np.float64, np.float64], float], k=3, threshold: float = 0.5
     ) -> None:
+        self.__k = k
+        self.__metric = metric
+        self.__threshold = threshold
+
         self.__change_points: list[int] = []
         self.__change_points_count = 0
 
-        self.__knncpd = KNNCPD(metric, k, threshold)
+        self.__knngraph: knngraph.KNNGraph | None = None
 
     def detect(self, window: Iterable[float | np.float64]) -> int:
         """Finds change points in window.
@@ -47,31 +52,85 @@ class KNNAlgorithm(Algorithm):
         self.__process_data(True, window)
         return self.__change_points.copy()
 
-    def __process_data(self, with_localization: bool, window: Iterable[float | np.float64]) -> None:
+    def __process_data(self, window: Iterable[float | np.float64]) -> None:
         """
         Processes a window of data to detect/localize all change points depending on working mode.
-        :param with_localization: boolean flag representing whether function needs to localize a change point.
         :param window: part of global data for change points analysis.
         """
-        self.__prepare()
-
         sample = deque(window)
         sample_size = len(sample)
         if sample_size == 0:
             return
 
-        self.__knncpd.process_sample(window)
-
-        return self.__knncpd.change_points if with_localization else self.__knncpd.change_points_count
-        # build a graph and examine each of the points in sample
-        # If the window size will be managed outside algorithm, algorithm should get
-        # whole window and build a graph on the whole window not point by point.
-        # I should make graph code more efficient. But probably first i should integrate it and then if i would have a time to optimize it.
-
-    def __prepare(self) -> None:
-        """
-        Clear algorithm's state (including change points and time related information) before data processing.
-        :param sample_size: an overall size of the sample.
-        """
-        self.__change_points = []
+        # Preparing.
+        self.__change_points: list[int] = []
         self.__change_points_count = 0
+
+        # Building graph.
+        self.__knngraph = knngraph.KNNGraph(window, self.__metric, self.__k)
+        self.__knngraph.build()
+
+        # Examining each point.
+        # Boundaries are always change points.
+        for time in range(1, len(window) - 1):
+            statistics = self.__calculate_statistics_in_point(time, len(window))
+
+            if self.__check_change_point(statistics):
+                self.__change_points.append(time)
+                self.__change_points_count += 1
+
+    def __calculate_statistics_in_point(self, time: int, window_size: int) -> float:
+        """
+        Calculate the statistics of the KNN graph in specified point.
+        """
+        assert self.__knngraph is not None, "Graph should not be None."
+
+        k = self.__k
+        n = window_size
+        n_1 = time
+        n_2 = n - time
+
+        h = 4 * (n_1 - 1) * (n_2 - 1) / ((n - 2) * (n - 3))
+
+        sum_1 = (1 / n) * sum(self.__knngraph.check_for_neighbourhood(i, j) * self.__knngraph.check_for_neighbourhood(j, i)
+                     for i in range(window_size) for j in range(window_size))
+        
+        sum_2 = (1 / n) * sum(self.__knngraph.check_for_neighbourhood(j, i) * self.__knngraph.check_for_neighbourhood(l, i)
+                     for i in range(window_size)
+                       for j in range(window_size)
+                         for l in range (window_size))
+
+        expectation = 4 * k * n_1 * (n_2) / (n - 1)
+        variance = (expectation / k) * (h * (sum_1 + k - (2 * k**2 / (n-1))) + (1 - h) * (sum_2 - k**2))
+        deviation = sqrt(variance)
+
+        permutation: np.array = np.arange(window_size)
+        np.random.shuffle(permutation)
+
+        statistics = -(self.__calculate_random_variable(permutation, time) - expectation) / deviation
+
+        return statistics
+
+    def __check_change_point(self, statistics) -> bool:
+        """
+        Check if change point occurs in current sequence.
+        :return: True if change point occurs, False otherwise.
+        """
+        return statistics > self.__threshold
+
+    def __calculate_random_variable(self, permutation: np.array, t: int, window_size: int) -> int:
+        """
+        Calculate a random variable from a permutation and a fixed point.
+        :param permutation: permutation of observations.
+        :param t: fixed point that splits the permutation.
+        :return: value of the random variable.
+        """
+        def b(i: int, j: int) -> bool:
+            pi = permutation[i]
+            pj = permutation[j]
+            return (pi <= t < pj) or (pj <= t < pi)
+
+        s = sum((self.__knngraph.check_for_neighbourhood(i, j) + self.__knngraph.check_for_neighbourhood(j, i)) * b(i, j)
+                     for i in range(window_size) for j in range(window_size))
+
+        return s
